+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- ConvertTo-PnPClientSidePage: Added support for keeping the source page Author, Editor, Created and Modified page properties (only when source page lives in SPO) (KeepPageCreationModificationInformation parameter)
+- ConvertTo-PnPClientSidePage: Added support for posting the created page as news (PostAsNews parameter)
+
+### Changed
+
 ## [3.13.1909.0]
 
 ### Added